supression de classe BackupManagerJGit
