+++ src/main/java/org/jabref/gui/autosaveandbackup/BackupManager.java
@@ -1,65 +1,46 @@
 package org.jabref.gui.autosaveandbackup;
 
+import java.io.File;
+import java.io.FileDescriptor;
+import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.Writer;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.StandardCopyOption;
-import java.nio.file.attribute.FileTime;
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
-import java.util.Queue;
 import java.util.Set;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
-import javafx.scene.control.TableColumn;
-
 import org.jabref.gui.LibraryTab;
-import org.jabref.gui.maintable.BibEntryTableViewModel;
-import org.jabref.gui.maintable.columns.MainTableColumn;
-import org.jabref.logic.bibtex.InvalidFieldValueException;
-import org.jabref.logic.exporter.AtomicFileWriter;
-import org.jabref.logic.exporter.BibWriter;
-import org.jabref.logic.exporter.BibtexDatabaseWriter;
-import org.jabref.logic.exporter.SelfContainedSaveConfiguration;
 import org.jabref.logic.preferences.CliPreferences;
-import org.jabref.logic.util.BackupFileType;
 import org.jabref.logic.util.CoarseChangeFilter;
-import org.jabref.logic.util.io.BackupFileUtil;
-import org.jabref.model.database.BibDatabase;
 import org.jabref.model.database.BibDatabaseContext;
-import org.jabref.model.database.event.BibDatabaseContextChangedEvent;
-import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibEntryTypesManager;
-import org.jabref.model.metadata.SaveOrder;
-import org.jabref.model.metadata.SelfContainedSaveOrder;
 
-import com.google.common.eventbus.Subscribe;
+import org.eclipse.jgit.api.Git;
+import org.eclipse.jgit.api.errors.GitAPIException;
+import org.eclipse.jgit.diff.DiffEntry;
+import org.eclipse.jgit.diff.DiffFormatter;
+import org.eclipse.jgit.lib.ObjectId;
+import org.eclipse.jgit.lib.ObjectLoader;
+import org.eclipse.jgit.lib.Repository;
+import org.eclipse.jgit.revwalk.RevCommit;
+import org.eclipse.jgit.revwalk.RevWalk;
+import org.eclipse.jgit.storage.file.FileRepositoryBuilder;
+import org.eclipse.jgit.treewalk.TreeWalk;
+import org.eclipse.jgit.treewalk.filter.PathFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/**
- * Backups the given bib database file from {@link BibDatabaseContext} on every {@link BibDatabaseContextChangedEvent}.
- * An intelligent {@link ExecutorService} with a {@link BlockingQueue} prevents a high load while making backups and
- * rejects all redundant backup tasks. This class does not manage the .bak file which is created when opening a
- * database.
- */
 public class BackupManager {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(BackupManager.class);
 
-    private static final int MAXIMUM_BACKUP_FILE_COUNT = 10;
-
     private static final int DELAY_BETWEEN_BACKUP_ATTEMPTS_IN_SECONDS = 19;
 
-    private static Set<BackupManager> runningInstances = new HashSet<>();
+    private static Set<BackupManager> runningInstances = new HashSet<BackupManager>();
 
     private final BibDatabaseContext bibDatabaseContext;
     private final CliPreferences preferences;
@@ -67,13 +48,11 @@ public class BackupManager {
     private final CoarseChangeFilter changeFilter;
     private final BibEntryTypesManager entryTypesManager;
     private final LibraryTab libraryTab;
+    private final Git git;
 
-    // Contains a list of all backup paths
-    // During writing, the less recent backup file is deleted
-    private final Queue<Path> backupFilesQueue = new LinkedBlockingQueue<>();
     private boolean needsBackup = false;
 
-    BackupManager(LibraryTab libraryTab, BibDatabaseContext bibDatabaseContext, BibEntryTypesManager entryTypesManager, CliPreferences preferences) {
+    BackupManager(LibraryTab libraryTab, BibDatabaseContext bibDatabaseContext, BibEntryTypesManager entryTypesManager, CliPreferences preferences) throws IOException, GitAPIException {
         this.bibDatabaseContext = bibDatabaseContext;
         this.entryTypesManager = entryTypesManager;
         this.preferences = preferences;
@@ -82,312 +61,244 @@ public class BackupManager {
 
         changeFilter = new CoarseChangeFilter(bibDatabaseContext);
         changeFilter.registerListener(this);
+
+        // Initialize Git repository
+        FileRepositoryBuilder builder = new FileRepositoryBuilder();
+        git = new Git(builder.setGitDir(new File(preferences.getFilePreferences().getBackupDirectory().toFile(), ".git"))
+                             .readEnvironment()
+                             .findGitDir()
+                             .build());
+        if (git.getRepository().getObjectDatabase().exists()) {
+            LOGGER.info("Git repository already exists");
+        } else {
+            git.init().call();
+            LOGGER.info("Initialized new Git repository");
+        }
     }
 
-    /**
-     * Determines the most recent backup file name
-     */
-    static Path getBackupPathForNewBackup(Path originalPath, Path backupDir) {
-        return BackupFileUtil.getPathForNewBackupFileAndCreateDirectory(originalPath, BackupFileType.BACKUP, backupDir);
+    public BackupManager start(LibraryTab libraryTab, BibDatabaseContext bibDatabaseContext, BibEntryTypesManager entryTypesManager, CliPreferences preferences, Path originalPath) throws IOException, GitAPIException {
+        BackupManager BackupManager = new BackupManager(libraryTab, bibDatabaseContext, entryTypesManager, preferences);
+        BackupManager.startBackupTask(preferences.getFilePreferences().getBackupDirectory(), originalPath);
+        runningInstances.add(BackupManager);
+        return BackupManager;
     }
 
-    /**
-     * Determines the most recent existing backup file name
-     */
-    static Optional<Path> getLatestBackupPath(Path originalPath, Path backupDir) {
-        return BackupFileUtil.getPathOfLatestExistingBackupFile(originalPath, BackupFileType.BACKUP, backupDir);
-    }
-
-    /**
-     * Starts the BackupManager which is associated with the given {@link BibDatabaseContext}. As long as no database
-     * file is present in {@link BibDatabaseContext}, the {@link BackupManager} will do nothing.
-     *
-     * This method is not thread-safe. The caller has to ensure that this method is not called in parallel.
-     *
-     * @param bibDatabaseContext Associated {@link BibDatabaseContext}
-     */
-    public static BackupManager start(LibraryTab libraryTab, BibDatabaseContext bibDatabaseContext, BibEntryTypesManager entryTypesManager, CliPreferences preferences) {
-        BackupManager backupManager = new BackupManager(libraryTab, bibDatabaseContext, entryTypesManager, preferences);
-        backupManager.startBackupTask(preferences.getFilePreferences().getBackupDirectory());
-        runningInstances.add(backupManager);
-        return backupManager;
-    }
-
-    /**
-     * Marks the backup as discarded at the library which is associated with the given {@link BibDatabaseContext}.
-     *
-     * @param bibDatabaseContext Associated {@link BibDatabaseContext}
-     */
-    public static void discardBackup(BibDatabaseContext bibDatabaseContext, Path backupDir) {
-        runningInstances.stream().filter(instance -> instance.bibDatabaseContext == bibDatabaseContext).forEach(backupManager -> backupManager.discardBackup(backupDir));
-    }
-
-    /**
-     * Shuts down the BackupManager which is associated with the given {@link BibDatabaseContext}.
-     *
-     * @param bibDatabaseContext Associated {@link BibDatabaseContext}
-     * @param createBackup True, if a backup should be created
-     * @param backupDir The path to the backup directory
-     */
-    public static void shutdown(BibDatabaseContext bibDatabaseContext, Path backupDir, boolean createBackup) {
-        runningInstances.stream().filter(instance -> instance.bibDatabaseContext == bibDatabaseContext).forEach(backupManager -> backupManager.shutdown(backupDir, createBackup));
+    @SuppressWarnings({"checkstyle:NoWhitespaceBefore", "checkstyle:WhitespaceAfter"})
+    public static void shutdown(BibDatabaseContext bibDatabaseContext, Path backupDir, boolean createBackup, Path originalPath) {
+        runningInstances.stream().filter(instance -> instance.bibDatabaseContext == bibDatabaseContext).forEach(backupManager -> backupManager.shutdownJGit(backupDir, createBackup, originalPath));
         runningInstances.removeIf(instance -> instance.bibDatabaseContext == bibDatabaseContext);
     }
 
-    /**
-     * Checks whether a backup file exists for the given database file. If it exists, it is checked whether it is
-     * newer and different from the original.
-     *
-     * In case a discarded file is present, the method also returns <code>false</code>, See also {@link #discardBackup(Path)}.
-     *
-     * @param originalPath Path to the file a backup should be checked for. Example: jabref.bib.
-     *
-     * @return <code>true</code> if backup file exists AND differs from originalPath. <code>false</code> is the
-     * "default" return value in the good case. In case a discarded file exists, <code>false</code> is returned, too.
-     * In the case of an exception <code>true</code> is returned to ensure that the user checks the output.
-     */
-    public static boolean backupFileDiffers(Path originalPath, Path backupDir) {
-        Path discardedFile = determineDiscardedFile(originalPath, backupDir);
-        if (Files.exists(discardedFile)) {
-            try {
-                Files.delete(discardedFile);
-            } catch (IOException e) {
-                LOGGER.error("Could not remove discarded file {}", discardedFile, e);
-                return true;
-            }
-            return false;
-        }
-        return getLatestBackupPath(originalPath, backupDir).map(latestBackupPath -> {
-            FileTime latestBackupFileLastModifiedTime;
-            try {
-                latestBackupFileLastModifiedTime = Files.getLastModifiedTime(latestBackupPath);
-            } catch (IOException e) {
-                LOGGER.debug("Could not get timestamp of backup file {}", latestBackupPath, e);
-                // If we cannot get the timestamp, we do show any warning
-                return false;
-            }
-            FileTime currentFileLastModifiedTime;
-            try {
-                currentFileLastModifiedTime = Files.getLastModifiedTime(originalPath);
-            } catch (IOException e) {
-                LOGGER.debug("Could not get timestamp of current file file {}", originalPath, e);
-                // If we cannot get the timestamp, we do show any warning
-                return false;
-            }
-            if (latestBackupFileLastModifiedTime.compareTo(currentFileLastModifiedTime) <= 0) {
-                // Backup is older than current file
-                // We treat the backup as non-different (even if it could differ)
-                return false;
-            }
-            try {
-                boolean result = Files.mismatch(originalPath, latestBackupPath) != -1L;
-                if (result) {
-                    LOGGER.info("Backup file {} differs from current file {}", latestBackupPath, originalPath);
-                }
-                return result;
-            } catch (IOException e) {
-                LOGGER.debug("Could not compare original file and backup file.", e);
-                // User has to investigate in this case
-                return true;
-            }
-        }).orElse(false);
+    @SuppressWarnings({"checkstyle:WhitespaceAfter", "checkstyle:LeftCurly"})
+    private void startBackupTask(Path backupDir, Path originalPath) {
+        executor.scheduleAtFixedRate(
+                () -> {
+                    try {
+                        performBackup(backupDir, originalPath);
+                    } catch (
+                            IOException |
+                            GitAPIException e) {
+                        LOGGER.error("Error during backup", e);
+                    }
+                },
+                DELAY_BETWEEN_BACKUP_ATTEMPTS_IN_SECONDS,
+                DELAY_BETWEEN_BACKUP_ATTEMPTS_IN_SECONDS,
+                TimeUnit.SECONDS);
     }
 
-    /**
-     * Restores the backup file by copying and overwriting the original one.
-     *
-     * @param originalPath Path to the file which should be equalized to the backup file.
-     */
-    public static void restoreBackup(Path originalPath, Path backupDir) {
-        Optional<Path> backupPath = getLatestBackupPath(originalPath, backupDir);
-        if (backupPath.isEmpty()) {
-            LOGGER.error("There is no backup file");
-            return;
-        }
-        try {
-            Files.copy(backupPath.get(), originalPath, StandardCopyOption.REPLACE_EXISTING);
-        } catch (IOException e) {
-            LOGGER.error("Error while restoring the backup file.", e);
-        }
-    }
-
-    Optional<Path> determineBackupPathForNewBackup(Path backupDir) {
-        return bibDatabaseContext.getDatabasePath().map(path -> BackupManager.getBackupPathForNewBackup(path, backupDir));
-    }
-
-    /**
-     * This method is called as soon as the scheduler says: "Do the backup"
-     *
-     * <em>SIDE EFFECT: Deletes oldest backup file</em>
-     *
-     * @param backupPath the full path to the file where the library should be backed up to
-     */
-    void performBackup(Path backupPath) {
+    private void performBackup(Path backupDir, Path originalPath) throws IOException, GitAPIException {
+        /*
+        needsBackup must be initialized
+         */
+        needsBackup = BackupManager.backupGitDiffers(backupDir, originalPath);
         if (!needsBackup) {
             return;
         }
 
-        // We opted for "while" to delete backups in case there are more than 10
-        while (backupFilesQueue.size() >= MAXIMUM_BACKUP_FILE_COUNT) {
-            Path oldestBackupFile = backupFilesQueue.poll();
-            try {
-                Files.delete(oldestBackupFile);
-            } catch (IOException e) {
-                LOGGER.error("Could not delete backup file {}", oldestBackupFile, e);
-            }
-        }
-        //l'ordre dans lequel les entrées BibTeX doivent être écrites dans le fichier de sauvegarde.
-        // Si l'utilisateur a trié la table d'affichage des entrées dans JabRef, cet ordre est récupéré.
-        // Sinon, un ordre par défaut est utilisé.
-        // code similar to org.jabref.gui.exporter.SaveDatabaseAction.saveDatabase
-        SelfContainedSaveOrder saveOrder = bibDatabaseContext
-                .getMetaData().getSaveOrder()
-                .map(so -> {
-                    if (so.getOrderType() == SaveOrder.OrderType.TABLE) {
-                        // We need to "flatten out" SaveOrder.OrderType.TABLE as BibWriter does not have access to preferences
-                        List<TableColumn<BibEntryTableViewModel, ?>> sortOrder = libraryTab.getMainTable().getSortOrder();
-                        return new SelfContainedSaveOrder(
-                                SaveOrder.OrderType.SPECIFIED,
-                                sortOrder.stream()
-                                         .filter(col -> col instanceof MainTableColumn<?>)
-                                         .map(column -> ((MainTableColumn<?>) column).getModel())
-                                         .flatMap(model -> model.getSortCriteria().stream())
-                                         .toList());
-                    } else {
-                        return SelfContainedSaveOrder.of(so);
-                    }
-                })
-                .orElse(SaveOrder.getDefaultSaveOrder());
+        // Add and commit changes
+        git.add().addFilepattern(".").call();
+        RevCommit commit = git.commit().setMessage("Backup at " + System.currentTimeMillis()).call();
+        LOGGER.info("Committed backup: {}", commit.getId());
 
-        //Elle configure la sauvegarde, en indiquant qu'aucune sauvegarde supplémentaire (backup) ne doit être créée,
-        // que l'ordre de sauvegarde doit être celui défini, et que les entrées doivent être formatées selon les préférences
-        // utilisateur.
-        SelfContainedSaveConfiguration saveConfiguration = (SelfContainedSaveConfiguration) new SelfContainedSaveConfiguration()
-                .withMakeBackup(false)
-                .withSaveOrder(saveOrder)
-                .withReformatOnSave(preferences.getLibraryPreferences().shouldAlwaysReformatOnSave());
-
-        // "Clone" the database context
-        // We "know" that "only" the BibEntries might be changed during writing (see [org.jabref.logic.exporter.BibDatabaseWriter.savePartOfDatabase])
-        //Chaque entrée BibTeX (comme un article, livre, etc.) est clonée en utilisant la méthode clone().
-        // Cela garantit que les modifications faites pendant la sauvegarde n'affecteront pas l'entrée originale.
-        List<BibEntry> list = bibDatabaseContext.getDatabase().getEntries().stream()
-                                                .map(BibEntry::clone)
-                                                .map(BibEntry.class::cast)
-                                                .toList();
-        BibDatabase bibDatabaseClone = new BibDatabase(list);
-        BibDatabaseContext bibDatabaseContextClone = new BibDatabaseContext(bibDatabaseClone, bibDatabaseContext.getMetaData());
-        //Elle définit l'encodage à utiliser pour écrire le fichier. Cela garantit que les caractères spéciaux sont bien sauvegardés.
-        Charset encoding = bibDatabaseContext.getMetaData().getEncoding().orElse(StandardCharsets.UTF_8);
-        // We want to have successful backups only
-        // Thus, we do not use a plain "FileWriter", but the "AtomicFileWriter"
-        // Example: What happens if one hard powers off the machine (or kills the jabref process) during writing of the backup?
-        //          This MUST NOT create a broken backup file that then jabref wants to "restore" from?
-        try (Writer writer = new AtomicFileWriter(backupPath, encoding, false)) {
-            BibWriter bibWriter = new BibWriter(writer, bibDatabaseContext.getDatabase().getNewLineSeparator());
-            new BibtexDatabaseWriter(
-                    bibWriter,
-                    saveConfiguration,
-                    preferences.getFieldPreferences(),
-                    preferences.getCitationKeyPatternPreferences(),
-                    entryTypesManager)
-                    // we save the clone to prevent the original database (and thus the UI) from being changed
-                    .saveDatabase(bibDatabaseContextClone);
-            backupFilesQueue.add(backupPath);
-
-            // We wrote the file successfully
-            // Thus, we currently do not need any new backup
-            this.needsBackup = false;
-        } catch (IOException e) {
-            logIfCritical(backupPath, e);
-        }
+        // Reset the backup flag
+        this.needsBackup = false;
     }
 
-    private static Path determineDiscardedFile(Path file, Path backupDir) {
-        return backupDir.resolve(BackupFileUtil.getUniqueFilePrefix(file) + "--" + file.getFileName() + "--discarded");
-    }
-
-    /**
-     * Marks the backups as discarded.
-     *
-     * We do not delete any files, because the user might want to recover old backup files.
-     * Therefore, we mark discarded backups by a --discarded file.
-     */
-    public void discardBackup(Path backupDir) {
-        Path path = determineDiscardedFile(bibDatabaseContext.getDatabasePath().get(), backupDir);
+    @SuppressWarnings("checkstyle:TodoComment")
+    public static void restoreBackup(Path originalPath, Path backupDir, ObjectId objectId) {
         try {
-            Files.createFile(path);
-        } catch (IOException e) {
-            LOGGER.info("Could not create backup file {}", path, e);
+            Git git = Git.open(backupDir.toFile());
+
+            git.checkout().setStartPoint(objectId.getName()).setAllPaths(true).call();
+            // Add commits to staging Area
+            git.add().addFilepattern(".").call();
+
+            // Commit with a message
+            git.commit().setMessage("Restored content from commit: " + objectId.getName()).call();
+
+            LOGGER.info("Restored backup from Git repository and committed the changes");
+        } catch (
+                IOException |
+                GitAPIException e) {
+            LOGGER.error("Error while restoring the backup", e);
         }
     }
 
-    private void logIfCritical(Path backupPath, IOException e) {
-        Throwable innermostCause = e;
-        while (innermostCause.getCause() != null) {
-            innermostCause = innermostCause.getCause();
-        }
-        boolean isErrorInField = innermostCause instanceof InvalidFieldValueException;
+    public static void restoreBackupj(Path originalPath, Path backupDir, ObjectId objectId) {
+        try {
 
-        // do not print errors in field values into the log during autosave
-        if (!isErrorInField) {
-            LOGGER.error("Error while saving to file {}", backupPath, e);
-        }
-    }
-
-    @Subscribe
-    public synchronized void listen(@SuppressWarnings("unused") BibDatabaseContextChangedEvent event) {
-        if (!event.isFilteredOut()) {
-            this.needsBackup = true;
+            Git git = Git.open(backupDir.toFile());
+            git.checkout().setName(objectId.getName()).call();
+            /*
+            faut une methode pour evite le branch nouveau
+            need a method to avoid the new branch
+             */
+            LOGGER.info("Restored backup from Git repository");
+        } catch (
+                IOException |
+                GitAPIException e) {
+            LOGGER.error("Error while restoring the backup", e);
         }
     }
+    /*
+        compare what is in originalPath and last commit
+        */
 
-    private void startBackupTask(Path backupDir) {
-        fillQueue(backupDir);//remplie backupFilesQueue les files .sav de le meme bibl
+    public static boolean backupGitDiffers(Path originalPath, Path backupDir) throws IOException, GitAPIException {
 
-        executor.scheduleAtFixedRate(
-                                     // We need to determine the backup path on each action, because we use the timestamp in the filename
-                                     () -> determineBackupPathForNewBackup(backupDir).ifPresent(path -> this.performBackup(path)),
-                                     DELAY_BETWEEN_BACKUP_ATTEMPTS_IN_SECONDS,
-                                     DELAY_BETWEEN_BACKUP_ATTEMPTS_IN_SECONDS,
-                                     TimeUnit.SECONDS);
-    }
-//La méthode fillQueue(backupDir) est définie dans le code et son rôle est de lister et d'ajouter
-// les fichiers de sauvegarde existants dans une file d'attente,
-    private void fillQueue(Path backupDir) {
-        if (!Files.exists(backupDir)) {
-            return;
-        }
-        bibDatabaseContext.getDatabasePath().ifPresent(databasePath -> {
-            // code similar to {@link org.jabref.logic.util.io.BackupFileUtil.getPathOfLatestExisingBackupFile}
-            final String prefix = BackupFileUtil.getUniqueFilePrefix(databasePath) + "--" + databasePath.getFileName();
-            try {
-                List<Path> allSavFiles = Files.list(backupDir)
-                                              // just list the .sav belonging to the given targetFile
-                                              .filter(p -> p.getFileName().toString().startsWith(prefix))//tous les files .sav commencerait par ce prefix
-                                              .sorted().toList();
-                backupFilesQueue.addAll(allSavFiles);
-            } catch (IOException e) {
-                LOGGER.error("Could not determine most recent file", e);
+        File repoDir = backupDir.toFile();
+        Repository repository = new FileRepositoryBuilder()
+                .setGitDir(new File(repoDir, ".git"))
+                .build();
+        try (Git git = new Git(repository)) {
+            ObjectId headCommitId = repository.resolve("HEAD"); // to get the latest commit id
+            if (headCommitId == null) {
+                // No commits in the repository, so there's no previous backup
+                return false;
             }
-        });
+            git.add().addFilepattern(originalPath.getFileName().toString()).call();
+            String relativePath = backupDir.relativize(originalPath).toString();
+            List<DiffEntry> diffs = git.diff()
+                                       .setPathFilter(PathFilter.create(relativePath)) // Utiliser PathFilter ici
+                                       .call();
+            return !diffs.isEmpty();
+        }
     }
 
-    /**
-     * Unregisters the BackupManager from the eventBus of {@link BibDatabaseContext}.
-     * This method should only be used when closing a database/JabRef in a normal way.
-     *
-     * @param backupDir The backup directory
-     * @param createBackup If the backup manager should still perform a backup
-     */
-    private void shutdown(Path backupDir, boolean createBackup) {
+    public List<DiffEntry> showDiffers(Path originalPath, Path backupDir, String CommitId) throws IOException, GitAPIException {
+
+        File repoDir = backupDir.toFile();
+        Repository repository = new FileRepositoryBuilder()
+                .setGitDir(new File(repoDir, ".git"))
+                .build();
+        /*
+        need a class to show the last ten backups indicating: date/ size/ number of entries
+         */
+
+        ObjectId oldCommit = repository.resolve(CommitId);
+        ObjectId newCommit = repository.resolve("HEAD");
+
+        FileOutputStream fos = new FileOutputStream(FileDescriptor.out);
+        DiffFormatter diffFr = new DiffFormatter(fos);
+        diffFr.setRepository(repository);
+        return diffFr.scan(oldCommit, newCommit);
+    }
+
+    // n is a counter incrementing by 1 when the user asks to see older versions (packs of 10)
+// and decrements by 1 when the user asks to see the pack of the 10 earlier versions
+// the scroll down: n->n+1 ; the scroll up: n->n-1
+    public List<RevCommit> retreiveCommits(Path backupDir, int n) throws IOException, GitAPIException {
+        List<RevCommit> retrievedCommits = new ArrayList<>();
+        // Open Git depository
+        try (Repository repository = Git.open(backupDir.toFile()).getRepository()) {
+            //  Use RevWalk to go through all commits
+            try (RevWalk revWalk = new RevWalk(repository)) {
+                // Start from HEAD
+                RevCommit startCommit = revWalk.parseCommit(repository.resolve("HEAD"));
+                revWalk.markStart(startCommit);
+
+                int count = 0;
+                int startIndex = n * 10;
+                int endIndex = startIndex + 9;
+
+                for (RevCommit commit : revWalk) {
+                    // Ignore commits before starting index
+                    if (count < startIndex) {
+                        count++;
+                        continue;
+                    }
+                    if (count >= endIndex) {
+                        break;
+                    }
+                    // Add commits to the main list
+                    retrievedCommits.add(commit);
+                    count++;
+                }
+            }
+        }
+
+        return retrievedCommits;
+    }
+
+    @SuppressWarnings("checkstyle:WhitespaceAround")
+    public List<List<String>> retrieveCommitDetails(List<RevCommit> commits, Path backupDir) throws IOException, GitAPIException {
+        List<List<String>> commitDetails;
+        try (Repository repository = Git.open(backupDir.toFile()).getRepository()) {
+            commitDetails = new ArrayList<>();
+
+            // Browse the list of commits given as a parameter
+            for (RevCommit commit : commits) {
+                // A list to stock details about the commit
+                List<String> commitInfo = new ArrayList<>();
+                commitInfo.add(commit.getName()); // ID of commit
+
+                // Get the size of files changes by the commit
+                try (TreeWalk treeWalk = new TreeWalk(repository)) {
+                    treeWalk.addTree(commit.getTree());
+                    treeWalk.setRecursive(true);
+                    long totalSize = 0;
+
+                    while (treeWalk.next()) {
+                        ObjectLoader loader = repository.open(treeWalk.getObjectId(0));
+                        totalSize += loader.getSize(); // size in bytes
+                    }
+
+                    // Convert the size to Kb or Mb
+                    String sizeFormatted = (totalSize > 1024 * 1024)
+                            ? String.format("%.2f Mo", totalSize / (1024.0 * 1024.0))
+                            : String.format("%.2f Ko", totalSize / 1024.0);
+
+                    commitInfo.add(sizeFormatted); // Add Formatted size
+                }
+
+                // adding date detail
+                Date date = commit.getAuthorIdent().getWhen();
+                commitInfo.add(date.toString());
+                // Add list of details to the main list
+                commitDetails.add(commitInfo);
+            }
+        }
+
+        return commitDetails;
+    }
+
+    private void shutdownJGit(Path backupDir, boolean createBackup, Path originalPath) {
         changeFilter.unregisterListener(this);
         changeFilter.shutdown();
         executor.shutdown();
 
         if (createBackup) {
-            // Ensure that backup is a recent one
-            determineBackupPathForNewBackup(backupDir).ifPresent(this::performBackup);
+            try {
+                performBackup(backupDir, originalPath);
+            } catch (IOException | GitAPIException e) {
+                LOGGER.error("Error during shutdown backup", e);
+            }
         }
     }
 }
+
+
+
+
+
+
